# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  admin: 'admin',
  user: 'user',
} as const;

export type UserProfileApprovalStatus = typeof UserProfileApprovalStatus[keyof typeof UserProfileApprovalStatus];


export const UserProfileApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UserProfile {
  id: number;
  clerkId?: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  picture?: string | null;
  role: UserProfileRole;
  approvalStatus: UserProfileApprovalStatus;
  /** @nullable */
  approvedUntil?: string | null;
  createdAt: string;
}

export type UserApprovalUpdateApprovalStatus = typeof UserApprovalUpdateApprovalStatus[keyof typeof UserApprovalUpdateApprovalStatus];


export const UserApprovalUpdateApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UserApprovalUpdate {
  approvalStatus: UserApprovalUpdateApprovalStatus;
  /** @nullable */
  durationDays?: number | null;
}

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  running: 'running',
  stopped: 'stopped',
  error: 'error',
  starting: 'starting',
  stopping: 'stopping',
} as const;

export type BotInstallStatus = typeof BotInstallStatus[keyof typeof BotInstallStatus];


export const BotInstallStatus = {
  none: 'none',
  installing: 'installing',
  ready: 'ready',
  failed: 'failed',
} as const;

export interface Bot {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: BotStatus;
  installStatus: BotInstallStatus;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  webhookUrl?: string | null;
  /** @nullable */
  token?: string | null;
  /** @nullable */
  uptimeSeconds?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface BotInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  phoneNumber?: string;
  webhookUrl?: string;
  token?: string;
}

export interface BotPatch {
  /** @minLength 1 */
  name?: string;
  description?: string;
  phoneNumber?: string;
  webhookUrl?: string;
  token?: string;
}

export type BotLogLevel = typeof BotLogLevel[keyof typeof BotLogLevel];


export const BotLogLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface BotLog {
  id: number;
  botId: number;
  level: BotLogLevel;
  message: string;
  createdAt: string;
}

export type BotLogInputLevel = typeof BotLogInputLevel[keyof typeof BotLogInputLevel];


export const BotLogInputLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface BotLogInput {
  level: BotLogInputLevel;
  message: string;
}

export interface ProcessInput {
  text: string;
}

export interface CommandResult {
  ok: boolean;
}

export interface DashboardStats {
  total: number;
  running: number;
  stopped: number;
  error: number;
  totalLogsToday: number;
}