# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhatsApp Bot Hosting Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: bots
  - name: logs
  - name: admin
  - name: dashboard

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated

  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden

  /admin/users/{id}:
    patch:
      operationId: updateUserApproval
      tags: [admin]
      summary: Approve or reject a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserApprovalUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
        "404":
          description: Not found

  /bots:
    get:
      operationId: listBots
      tags: [bots]
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bot"
    post:
      operationId: createBot
      tags: [bots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "400":
          description: Bad request

  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
    patch:
      operationId: updateBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotPatch"
      responses:
        "200":
          description: Bot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found
    delete:
      operationId: deleteBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /bots/{id}/start:
    post:
      operationId: startBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found

  /bots/{id}/stop:
    post:
      operationId: stopBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found

  /bots/{id}/restart:
    post:
      operationId: restartBot
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Not found

  /bots/{id}/logs:
    get:
      operationId: getBotLogs
      tags: [logs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BotLog"
        "404":
          description: Bot not found
    post:
      operationId: addBotLog
      tags: [logs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotLogInput"
      responses:
        "201":
          description: Log added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotLog"

  /bots/{id}/input:
    post:
      operationId: sendBotInput
      tags: [bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessInput"
      responses:
        "200":
          description: Input sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "404":
          description: Not found

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    UserProfile:
      type: object
      required: [id, email, role, approvalStatus, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        picture:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        approvedUntil:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time

    UserApprovalUpdate:
      type: object
      required: [approvalStatus]
      properties:
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        durationDays:
          type: ["integer", "null"]

    Bot:
      type: object
      required: [id, name, status, installStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [running, stopped, error, starting, stopping]
        installStatus:
          type: string
          enum: [none, installing, ready, failed]
        phoneNumber:
          type: ["string", "null"]
        webhookUrl:
          type: ["string", "null"]
        token:
          type: ["string", "null"]
        uptimeSeconds:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    BotInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        phoneNumber:
          type: string
        webhookUrl:
          type: string
        token:
          type: string

    BotPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        phoneNumber:
          type: string
        webhookUrl:
          type: string
        token:
          type: string

    BotLog:
      type: object
      required: [id, botId, level, message, createdAt]
      properties:
        id:
          type: integer
        botId:
          type: integer
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string
        createdAt:
          type: string
          format: date-time

    BotLogInput:
      type: object
      required: [level, message]
      properties:
        level:
          type: string
          enum: [info, warn, error, debug]
        message:
          type: string

    ProcessInput:
      type: object
      required: [text]
      properties:
        text:
          type: string

    CommandResult:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    DashboardStats:
      type: object
      required: [total, running, stopped, error, totalLogsToday]
      properties:
        total:
          type: integer
        running:
          type: integer
        stopped:
          type: integer
        error:
          type: integer
        totalLogsToday:
          type: integer