# lib/api-zod/src/generated/types/userApprovalUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserApprovalUpdateApprovalStatus } from './userApprovalUpdateApprovalStatus';

export interface UserApprovalUpdate {
  approvalStatus: UserApprovalUpdateApprovalStatus;
  /** @nullable */
  durationDays?: number | null;
}