# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileApprovalStatus } from './userProfileApprovalStatus';
import type { UserProfileRole } from './userProfileRole';

export interface UserProfile {
  id: number;
  clerkId?: string;
  email: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  picture?: string | null;
  role: UserProfileRole;
  approvalStatus: UserProfileApprovalStatus;
  /** @nullable */
  approvedUntil?: Date | null;
  createdAt: Date;
}